Extract Campaign Details in separate component
